+++ NEWS.md
@@ -1,5 +1,8 @@
 # reprex (development version)
 
+* `venue = "so"` (SO = Stack Overflow) has converged with default `venue = "gh"` (GitHub). As of January 2019, SO [supports CommonMark fenced code blocks](https://meta.stackexchange.com/questions/125148/implement-style-fenced-markdown-code-blocks/322000#322000). The only remaining difference is that Stack
+Overflow does not support the collapsible details tag that we use on GitHub for the session info requested via `si = TRUE` (#231).
+
 # reprex 0.3.0
 
 * The `crayon.enabled` option is explicitly set to `FALSE` when rendering the reprex (#238, #239).